requirements.txt
paho-mqtt>=1.6.1
fastapi>=0.100.0
uvicorn>=0.23.0
